\name{soilwaterfun-package}
\alias{soilwaterfun-package}
\alias{soilwaterfun}
\docType{package}
\title{
\packageTitle{soilwaterfun}
}
\description{
\packageDescription{soilwaterfun}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{soilwaterfun}
\packageIndices{soilwaterfun}
}
\author{
\packageAuthor{soilwaterfun}

Maintainer: \packageMaintainer{soilwaterfun}
}
\keyword{ package }
