\name{soilwaterfun-package}
\alias{soilwaterfun-package}
\alias{soilwaterfun}
\docType{package}
\title{Functions for soil water retention and soil hydraulic conductivity.}
\description{The package implements commonly used functions for soil water retention (water content = f(matrix potential)) and soil hydraulic conductivity (hydraulic conductivity K = f(matrix potential), or = f(water content)). The functions for for soil water retention implemented are: van Genuchten 1980; Brooks & Corey 1964; Campbell 1974. The functions for soil hydraulic conductivity implemented are: Mualem 1976; Brooks & Corey 1964; Campbell 1974. NB: The package soilwaterptf contains pedotransfer function for estimating the parameters of these functions from basic soil properties (valid for european soils).}
\details{
\tabular{ll}{Package: \tab soilwaterfun\cr
Version: \tab 1.0.9\cr
Date: \tab 2015-05-05\cr
Title: \tab Functions for soil water retention and soil hydraulic conductivity.\cr
Author: \tab Julien Moeys <jules_m78-soiltexture@yahoo.fr> [aut, cre]\cr
Authors@R: \tab c( person( "Julien", "Moeys", role = c("aut", "cre"), email = "jules_m78-soiltexture@yahoo.fr" ) )\cr
Maintainer: \tab Julien Moeys <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.1.1)\cr
License: \tab AGPL-3\cr
URL: \tab http://www.mark.slu.se, http://ckb.slu.se, http://julienmoeys.free.fr\cr}
}
\author{Julien Moeys <jules_m78-soiltexture@yahoo.fr> [aut, cre]}

\keyword{ package }


